[UPDATE] addtl work

+++ misc_optimize_images/main.go
@@ -5,6 +5,7 @@ import (
 	"os"
 	"path/filepath"
 	"regexp"
+	"runtime"
 	"sync"
 
 	"github.com/windmillcode/go_scripts/utils"
@@ -48,18 +49,18 @@ func main() {
 	for _, project := range projectsList {
 		app := filepath.Join(project)
 		normalizedBackupLocation := ""
-		// if runtime.GOOS == "windows"{
-		// 	normalizedBackupLocation = filepath.Join(backupLocation,utils.RemoveDrivePath(app))
-		// } else {
-		// }
-		normalizedBackupLocation = filepath.Join(backupLocation, app)
+		if runtime.GOOS == "windows"{
+			normalizedBackupLocation = filepath.Join(backupLocation,utils.RemoveDrivePath(app))
+		} else {
+			normalizedBackupLocation = filepath.Join(backupLocation, app)
+		}
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
 			utils.CopyDir(app, normalizedBackupLocation)
 			allEntries, err := utils.GetItemsInFolderRecursive(app, true)
 			if err != nil {
-				fmt.Println("An error occured while recursively goin through the directory", err)
+				fmt.Println("An error occured while recursively going through the directory", err)
 			}
 			for _, entry := range allEntries {
 				prefixImage := utils.HasSuffixInArray(entry, []string{".png", ".gif", ".ico", ".jpg",  ".webp", ".ico"}, true)
@@ -69,9 +70,14 @@ func main() {
 					destImage := utils.HasPrefixInArray(prefixImage,[]string{imageFolderPath+"\\"},true)
 
 					utils.RunCommandInSpecificDirectory("convert", []string{
-						"-quality",
-						optimizePercent, imageFile, fmt.Sprintf("%s%s", destImage, ".jpg")}, imageFolderPath)
-					os.Remove(entry)
+						"-quality", optimizePercent,
+						 imageFile,
+						 "-background", "#FFFFFF", "-flatten",
+						 fmt.Sprintf("%s%s", destImage, ".jpg")}, imageFolderPath)
+
+					if utils.HasSuffixInArray(entry, []string{".jpg"}, true) == ""{
+						os.Remove(entry)
+					}
 
 				}
 			}
